==> notebooks_stories/1_generate/06_story_viz.ipynb <==
%load_ext autoreload
%autoreload 2
import os
import matplotlib.pyplot as plt
import seaborn as sns
from os.path import join
from tqdm import tqdm
import pandas as pd
import sys
from IPython.display import display, HTML
from typing import List
import numpy as np
import matplotlib
import imodelsx.util
from copy import deepcopy
import re
# from sasc import notebook_helper
import scipy.special
from pprint import pprint
import joblib
import sasc.viz as viz
from sasc.config import RESULTS_DIR


# EXPT_NAME = 'relationships_mar9'
# EXPT_NAME = 'huth2016clusters_mar21_i_time_traveled'
# EXPT_NAME = 'voxels_mar21_hands_arms_emergency'
EXPT_NAME = f'uts03___jun14___seed=5'
EXPT_DIR = join(RESULTS_DIR, 'stories', 'default', EXPT_NAME)
# EXPT_DIRS = [join(RESULTS_DIR, 'stories', f'uts02_concepts_pilot_mar22_seed={seed}')
            #  for seed in range(1, 8)]
rows = joblib.load(join(EXPT_DIR, f'rows.pkl'))
expls = rows.expl.values
# paragraphs = rows.paragraphs.values
# prompts = rows.prompt.values
# voxel_nums = rows.module_num.values
# subjects = rows.subject.values

### Story visualization

**Explanation** <> Story match

scores_data_story = joblib.load(join(EXPT_DIR, f'scores_data_story.pkl'))
s_data = notebook_helper.viz_paragraphs(
    paragraphs, scores_data_story, expls, prompts,
    normalize_to_range=True, moving_average=True, shift_to_range=True)
display(HTML(s_data))

**Module** <> Story Match

s_mod = ''
scores_mod_story = [scores_mod_dict['scores_all'][i][i]
                    for i in range(len(scores_mod_dict['scores_all']))]
for i in range(len(paragraphs)):
    scores_i = scores_mod_story[i]

    # normalize to 0-1 range
    scores_i = (scores_i - scores_i.min()) / (scores_i.max() - scores_i.min())
    # scores_mod_i = scipy.special.softmax(scores_mod_i)
    # scores_i = mprompt.viz.moving_average(scores_i, n=3)
    scores_i = scores_i / 2 + 0.5  # shift to 0.5-1 range
    s_mod += ' ' + \
        mprompt.viz.colorize(paragraphs[i].split(
        ), scores_i, title=expls[i], subtitle=prompts[i])
# display(HTML(s_mod))

template_style = '''<style type="text/css">
#wrap {
   width:100%;
   margin:0 auto;
}
#left_col {
   float:left;
   width:48%;
}
#right_col {
   float:right;
   width:48%;
}
# body {
#     background-color:#b0c4de;
# }
</style>
'''

template = '''
<body>
    <div id="wrap">
        <div id="left_col">
            {col_left}
        </div>
        <div id="right_col">
            {col_right}
        </div>
    </div>
</body>
'''

display(HTML(template_style + template.format(col_left=s_data,
        col_right=s_mod)), metadata=dict(isolated=True))
